Create add collaborator form
- Set styles for Collaborator form
